requirements.txt
python-telegram-bot>=21.0.1
python-dotenv==1.0.0
gspread==5.12.4
google-auth-oauthlib==1.2.0